Added film history.

+++ src/Inventory/Application/UseCases/Command/UpdateFilm/UpdateFilmCommand.php
@@ -14,10 +14,10 @@ readonly class UpdateFilmCommand implements CommandInterface
      * @param int         $id       the ID of the object
      * @param string      $filmType the type of the film
      * @param string|null $name     the name of the object (optional)
-     * @param int|null    $length   the length of the film (optional)
+     * @param float|null    $length   the length of the film (optional)
      * @param string|null $type     the type of the object (optional)
      */
-    public function __construct(public int $id, public string $filmType, public ?string $name = null, public ?int $length = null, public ?string $type = null)
+    public function __construct(public int $id, public string $filmType, public ?string $name = null, public ?float $length = null, public ?string $type = null)
     {
     }
 }
