+++ src/Inventory/Application/UseCases/Query/DTO/FilmData.php
@@ -16,10 +16,10 @@ final readonly class FilmData
      *
      * @param int    $id     the unique identifier
      * @param string $name   the name of the object
-     * @param int    $length the length of the object
+     * @param float  $length the length of the object
      * @param string $type   the type of the object
      */
-    public function __construct(public int $id, public string $name, public int $length, public string $type)
+    public function __construct(public int $id, public string $name, public float $length, public string $type)
     {
     }
 }
