+++ src/Inventory/Application/UseCases/Query/DTO/RollData.php
@@ -15,10 +15,10 @@ final readonly class RollData
      * Constructor for a class.
      *
      * @param int $id     The ID of the object
-     * @param int $length The length of the object
+     * @param float $length The length of the object
      * @param int $filmId The film ID associated with the object
      */
-    public function __construct(public int $id, public int $length, public int $filmId)
+    public function __construct(public int $id, public float $length, public int $filmId)
     {
     }
 }
