+++ src/Orders/Domain/DTO/FilmData.php
@@ -14,10 +14,10 @@ final readonly class FilmData
      *
      * @param int    $id       the unique identifier of the entry
      * @param string $name     the name of the entry
-     * @param int    $length   the length of the entry
+     * @param float    $length   the length of the entry
      * @param string $filmType the roll type of the entry
      */
-    public function __construct(public int $id, public string $name, public int $length, public string $filmType)
+    public function __construct(public int $id, public string $name, public float $length, public string $filmType)
     {
     }
 }
